Changed the Database to track Users per Server, that way it allows for a User to be in multiple games.
Modified all the remaining SQL Functions to show this change.

Added the SERVERS table, to track the Servers the bot is in.

Added the onJoin listener to add the Server to the SERVERS table on Join.

Added the onLeave listener to remove the Server from the SERVERS table on Leave, and to delete all Users in said server on Leave, if a User is in multiple servers it will remain in their other servers, the nation in the leaving server is deleted however.

Created the Slash Commands (Started, not completed, WIP)

Passed the API to all commands, for further use. (WIP)

The createNation command now allows for attachments of the flag rather than URLs, but allows either one. The createNation command now can take "Nations with multiple words" in it. Current use of the command is (createNation @person URL[Optional] NAME) If the user wishes to add the flag via attachment it would be as follows (createNation @person NAME).